refactor: selection
